set for performance &battery saveù
