+++ arch/arm/configs/trebon_defconfig
@@ -570,24 +570,29 @@ CONFIG_CPU_FREQ_TABLE=y
 CONFIG_CPU_FREQ_STAT=y
 # CONFIG_CPU_FREQ_STAT_DETAILS is not set
 # CONFIG_CPU_FREQ_DEFAULT_GOV_PERFORMANCE is not set
-CONFIG_CPU_FREQ_DEFAULT_GOV_HYPER=y 
-# CONFIG_CPU_FREQ_DEFAULT_GOV_LIONHEART is not set
 # CONFIG_CPU_FREQ_DEFAULT_GOV_POWERSAVE is not set
 # CONFIG_CPU_FREQ_DEFAULT_GOV_USERSPACE is not set
 # CONFIG_CPU_FREQ_DEFAULT_GOV_ONDEMAND is not set
-# CONFIG_CPU_FREQ_DEFAULT_GOV_ONDEMANDX is not set
 # CONFIG_CPU_FREQ_DEFAULT_GOV_CONSERVATIVE is not set
 # CONFIG_CPU_FREQ_DEFAULT_GOV_INTERACTIVE is not set
+# CONFIG_CPU_FREQ_DEFAULT_GOV_LIONHEART is not set
+CONFIG_CPU_FREQ_DEFAULT_GOV_HYPER=y
 CONFIG_CPU_FREQ_GOV_PERFORMANCE=y
 CONFIG_CPU_FREQ_GOV_POWERSAVE=y
 CONFIG_CPU_FREQ_GOV_USERSPACE=y
 CONFIG_CPU_FREQ_GOV_ONDEMAND=y
-CONFIG_CPU_FREQ_GOV_HYPER=y
-CONFIG_CPU_FREQ_GOV_ONDEMANDX=y
 CONFIG_CPU_FREQ_GOV_INTELLIDEMAND2=y
-CONFIG_CPU_FREQ_GOV_LIONHEART=y
 CONFIG_CPU_FREQ_GOV_INTERACTIVE=y
 CONFIG_CPU_FREQ_GOV_CONSERVATIVE=y
+CONFIG_CPU_FREQ_GOV_ONDEMANDX=y
+CONFIG_CPU_FREQ_GOV_ONDEMANDX_INPUT=y
+CONFIG_CPU_FREQ_GOV_LIONHEART=y
+CONFIG_CPU_FREQ_GOV_HYPER=y
+CONFIG_CPU_FREQ_MIN_TICKS=10
+CONFIG_CPU_FREQ_SAMPLING_LATENCY_MULTIPLIER=1000
+# CONFIG_SEC_DVFS is not set
+CONFIG_SEC_LIMIT_LCD_OFF_CPU_MAX_FREQ=y
+# CONFIG_VDD_USERSPACE is not set
 # CONFIG_CPU_IDLE is not set
 CONFIG_CPU_FREQ_MSM=y
