+++ drivers/cpufreq/Kconfig
@@ -57,22 +57,7 @@ config CPU_FREQ_DEFAULT_GOV_PERFORMANCE
 	  Use the CPUFreq governor 'performance' as default. This sets
 	  the frequency statically to the highest frequency supported by
 	  the CPU.
-	  
-config CPU_FREQ_DEFAULT_GOV_HYPER
-	bool "hyper"
-	select CPU_FREQ_GOV_HYPER
-	---help---
-	  Use the CPUFreq governor 'hyper' as default. This allows
-	  you to get a full dynamic cpu frequency capable system by simply
-	  loading your cpufreq low-level hardware driver, using the
-	  'hyper' governor for latency-sensitive workloads.	  
-	  
-config CPU_FREQ_DEFAULT_GOV_LIONHEART
-	bool "lionheart"
-	select CPU_FREQ_GOV_LIONHEART
-	help
-	  Use the CPUFreq governor 'lionheart' as default.	  
-	  
+
 config CPU_FREQ_DEFAULT_GOV_POWERSAVE
 	bool "powersave"
 	depends on EXPERT
@@ -81,7 +66,7 @@ config CPU_FREQ_DEFAULT_GOV_POWERSAVE
 	  Use the CPUFreq governor 'powersave' as default. This sets
 	  the frequency statically to the lowest frequency supported by
 	  the CPU.
-	  
+
 config CPU_FREQ_DEFAULT_GOV_USERSPACE
 	bool "userspace"
 	select CPU_FREQ_GOV_USERSPACE
@@ -102,20 +87,6 @@ config CPU_FREQ_DEFAULT_GOV_ONDEMAND
 	  Be aware that not all cpufreq drivers support the ondemand
 	  governor. If unsure have a look at the help section of the
 	  driver. Fallback governor will be the performance governor.
-	  
-config CPU_FREQ_DEFAULT_GOV_ONDEMANDX
-	bool "ondemandx"
-	select CPU_FREQ_GOV_ONDEMANDX
-	select CPU_FREQ_GOV_PERFORMANCE
-	help
-	  Use the CPUFreq governor 'ondemand' as default. This allows
-	  you to get a full dynamic frequency capable system by simply
-	  loading your cpufreq low-level hardware driver.
-	  Be aware that not all cpufreq drivers support the ondemand
-	  governor. If unsure have a look at the help section of the
-	  driver. Fallback governor will be the performance governor.
-	  OndemandX has built in sleep profile, but not working Sysfs
-	  interface.	  
 
 config CPU_FREQ_DEFAULT_GOV_CONSERVATIVE
 	bool "conservative"
@@ -137,6 +108,15 @@ config CPU_FREQ_DEFAULT_GOV_INTERACTIVE
 	  you to get a full dynamic cpu frequency capable system by simply
 	  loading your cpufreq low-level hardware driver, using the
 	  'interactive' governor for latency-sensitive workloads.
+config CPU_FREQ_DEFAULT_GOV_LIONHEART
+        bool "lionheart"
+        select CPU_FREQ_GOV_LIONHEART
+        help
+         Use the CPUFreq governor 'lionheart' as default.
+config CPU_FREQ_DEFAULT_GOV_HYPER
+        bool "HYPER"
+        select CPU_FREQ_GOV_HYPER 
+
 endchoice
 
 config CPU_FREQ_GOV_PERFORMANCE
@@ -160,7 +140,7 @@ config CPU_FREQ_GOV_POWERSAVE
 	  module will be called cpufreq_powersave.
 
 	  If in doubt, say Y.
-	  
+
 config CPU_FREQ_GOV_USERSPACE
 	tristate "'userspace' governor for userspace frequency scaling"
 	help
@@ -193,33 +173,7 @@ config CPU_FREQ_GOV_ONDEMAND
 	  For details, take a look at linux/Documentation/cpu-freq.
 
 	  If in doubt, say N.
-	  
-config CPU_FREQ_GOV_HYPER
-	tristate "'hyper' cpufreq governor"
-	depends on CPU_FREQ
-	---help---
-	  'hyper' - A tweaked "ondemand" based smart and smooth optimized governor!
-	
-	   If in doubt, say N.	  
-	  
-config CPU_FREQ_GOV_ONDEMANDX
-	tristate "'ondemandx' cpufreq policy governor"
-	select CPU_FREQ_TABLE
-	help
-	  'ondemand' - This driver adds a dynamic cpufreq policy governor.
-	  The governor does a periodic polling and
-	  changes frequency based on the CPU utilization.
-	  The support for this governor depends on CPU capability to
-	  do fast frequency switching (i.e, very low latency frequency
-	  transitions).
 
-	  To compile this driver as a module, choose M here: the
-	  module will be called cpufreq_ondemand.
-
-	  For details, take a look at linux/Documentation/cpu-freq.
-
-	  If in doubt, say N.	  
-	  
 config CPU_FREQ_GOV_INTELLIDEMAND2
 	tristate "'Intellidemand2' cpufreq governor"
 	depends on CPU_FREQ
@@ -227,11 +181,6 @@ config CPU_FREQ_GOV_INTELLIDEMAND2
 	  'Intellidemand2' - a "smart" optimized governor for the hero.
 
 	  If in doubt, say N.  
-	                   
-config CPU_FREQ_GOV_LIONHEART
-	tristate "'lionheart' cpufreq governor"
-	depends on CPU_FREQ         
-         
 config CPU_FREQ_GOV_INTERACTIVE
 	tristate "'interactive' cpufreq policy governor"
 	help
@@ -271,6 +220,74 @@ config CPU_FREQ_GOV_CONSERVATIVE
 	  For details, take a look at linux/Documentation/cpu-freq.
 
 	  If in doubt, say N.
+
+config CPU_FREQ_GOV_ONDEMANDX
+        tristate "'ondemandx' cpufreq policy governor"
+        select CPU_FREQ_TABLE
+        help
+         'ondemand' - This driver adds a dynamic cpufreq policy governor.
+         The governor does a periodic polling and
+         changes frequency based on the CPU utilization.
+         The support for this governor depends on CPU capability to
+         do fast frequency switching (i.e, very low latency frequency
+         transitions).
+
+         To compile this driver as a module, choose M here: the
+         module will be called cpufreq_ondemand.
+
+         For details, take a look at linux/Documentation/cpu-freq.
+
+         If in doubt, say N.
+
+config CPU_FREQ_GOV_ONDEMANDX_INPUT
+        bool "Ramp up CPU frequency on input events"
+        default y
+        depends on CPU_FREQ_GOV_ONDEMANDX
+        help
+         Enable installation of an input event handler which will ramp up the
+         CPU to max frequency when an input event is received.
+config CPU_FREQ_GOV_LIONHEART
+        tristate "'lionheart' cpufreq governor"
+        depends on CPU_FREQ
+config CPU_FREQ_GOV_HYPER
+        tristate "'HYPER' cpufreq policy governor"
+        depends on CPU_FREQ
+
+config CPU_FREQ_MIN_TICKS
+        int "Ticks between governor polling interval."
+        default 10
+        help
+         Minimum number of ticks between polling interval for governors.
+
+config CPU_FREQ_SAMPLING_LATENCY_MULTIPLIER
+        int "Sampling rate multiplier for governors."
+        default 1000
+        help
+         Sampling latency rate multiplied by the cpu switch latency.
+         Affects governor polling.
+
+config SEC_DVFS
+        bool "DVFS job"
+        default n
+        depends on CPU_FREQ
+
+config SEC_DVFS_UNI
+        bool "DVFS job for Uni Processer"
+        default n
+        depends on SEC_DVFS
+
+config SEC_LIMIT_LCD_OFF_CPU_MAX_FREQ
+        bool
+        default y
+        depends on CPU_FREQ_GOV_ONDEMAND
+
+config VDD_USERSPACE
+  bool "VDD sysfs interface"
+  default n
+  depends on CPU_FREQ_STAT
+  help
+    exposes the VDD table to userspace
+    allows users to adjust voltages on the fly
 
 menu "x86 CPU frequency scaling drivers"
 depends on X86
