+++ drivers/cpufreq/cpufreq_HYPER.c
@@ -1,0 +1,1107 @@
+/*
+* drivers/cpufreq/cpufreq_HYPER.c
+*
+* Copyright (C) 2001 Russell King
+* (C) 2003 Venkatesh Pallipadi <venkatesh.pallipadi@intel.com>.
+* Jun Nakajima <jun.nakajima@intel.com>
+*
+* 2012 Minor Edits by Sar Castillo <sar.castillo@gmail.com>
+* 2012 MAR heavy addons by DORIMANX <yuri@bynet.co.il>
+* This program is free software; you can redistribute it and/or modify
+* it under the terms of the GNU General Public License version 2 as
+* published by the Free Software Foundation.
+*/
+
+#include <linux/kernel.h>
+#include <linux/module.h>
+#include <linux/init.h>
+#include <linux/cpufreq.h>
+#include <linux/cpu.h>
+#include <linux/jiffies.h>
+#include <linux/kernel_stat.h>
+#include <linux/mutex.h>
+#include <linux/hrtimer.h>
+#include <linux/tick.h>
+#include <linux/ktime.h>
+#include <linux/sched.h>
+
+#include <linux/input.h>
+#include <linux/workqueue.h>
+#include <linux/slab.h>
+
+/* if kernel support suspend freq changes then disble GOV ability to change suspend freq */
+//#define SUSPEND_FREQ_ON
+
+#ifdef CONFIG_HAS_EARLYSUSPEND
+#ifdef SUSPEND_FREQ_ON
+#include <linux/earlysuspend.h>
+#endif
+#endif
+
+/*
+* dbs is used in this file as a shortform for demandbased switching
+* It helps to keep variable names smaller, simpler
+*/
+
+#define DEF_FREQUENCY_DOWN_DIFFERENTIAL (25)
+#define MIN_FREQUENCY_DOWN_DIFFERENTIAL                (1)
+#define DEF_FREQUENCY_UP_THRESHOLD (70)
+#define DEF_SAMPLING_DOWN_FACTOR (2)
+#define MAX_SAMPLING_DOWN_FACTOR (100000)
+#define MICRO_FREQUENCY_DOWN_DIFFERENTIAL (3)
+#define MICRO_FREQUENCY_UP_THRESHOLD (50)
+#define MICRO_FREQUENCY_MIN_SAMPLE_RATE (10000)
+#define MIN_FREQUENCY_UP_THRESHOLD (11)
+#define MAX_FREQUENCY_UP_THRESHOLD (100)
+#define FREQ_STEP (35)
+#define UP_THRESHOLD_AT_MIN_FREQ (40)
+#define FREQ_FOR_RESPONSIVENESS (756000)
+
+#ifdef CONFIG_HAS_EARLYSUSPEND
+#ifdef SUSPEND_FREQ_ON
+#define DEF_SUSPEND_FREQ                        (756000)
+#define FREQ_STEP_SUSPEND (40)
+#define SAMPLING_FACTOR_SUSPEND                        (3)
+#define DEF_FREQUENCY_UP_THRESHOLD_SUSPEND        (95)
+#endif
+#endif
+
+/*
+* The polling frequency of this governor depends on the capability of
+* the processor. Default polling frequency is 1000 times the transition
+* latency of the processor. The governor will work on any processor with
+* transition latency <= 10mS, using appropriate sampling
+* rate.
+* For CPUs with transition latency > 10mS (mostly drivers with CPUFREQ_ETERNAL)
+* this governor will not work.
+* All times here are in uS.
+*/
+#define MIN_SAMPLING_RATE_RATIO                        (1)
+
+static unsigned int min_sampling_rate;
+
+#define LATENCY_MULTIPLIER                        (1000)
+#define MIN_LATENCY_MULTIPLIER                        (100)
+#define TRANSITION_LATENCY_LIMIT                (10 * 1000 * 1000)
+
+static void do_dbs_timer(struct work_struct *work);
+static int cpufreq_governor_dbs(struct cpufreq_policy *policy,
+                                unsigned int event);
+
+#ifndef CONFIG_CPU_FREQ_DEFAULT_GOV_HYPER
+static
+#endif
+struct cpufreq_governor cpufreq_gov_HYPER = {
+       .name = "HYPER",
+       .governor = cpufreq_governor_dbs,
+       .max_transition_latency = TRANSITION_LATENCY_LIMIT,
+       .owner = THIS_MODULE,
+};
+
+/* Sampling types */
+enum {DBS_NORMAL_SAMPLE, DBS_SUB_SAMPLE};
+
+struct cpu_dbs_info_s {
+        cputime64_t prev_cpu_idle;
+        cputime64_t prev_cpu_iowait;
+        cputime64_t prev_cpu_wall;
+        cputime64_t prev_cpu_nice;
+        struct cpufreq_policy *cur_policy;
+        struct delayed_work work;
+        struct cpufreq_frequency_table *freq_table;
+        unsigned int freq_lo;
+        unsigned int freq_lo_jiffies;
+        unsigned int freq_hi_jiffies;
+        unsigned int rate_mult;
+        int cpu;
+        unsigned int sample_type:1;
+        /*
+         * percpu mutex that serializes governor limit change with
+         * do_dbs_timer invocation. We do not want do_dbs_timer to run
+         * when user is changing the governor or limits.
+         */
+        struct mutex timer_mutex;
+};
+static DEFINE_PER_CPU(struct cpu_dbs_info_s, od_cpu_dbs_info);
+
+static unsigned int dbs_enable;        /* number of CPUs using this policy */
+
+/*
+* dbs_mutex protects dbs_enable in governor start/stop.
+*/
+static DEFINE_MUTEX(dbs_mutex);
+
+static struct dbs_tuners {
+        unsigned int sampling_rate;
+        unsigned int up_threshold;
+        unsigned int up_threshold_min_freq;
+        unsigned int down_differential;
+        unsigned int ignore_nice;
+        unsigned int sampling_down_factor;
+        unsigned int powersave_bias;
+        unsigned int io_is_busy;
+        unsigned int freq_step;
+        unsigned int freq_responsiveness;
+        unsigned int suspend_freq;
+#ifdef CONFIG_HAS_EARLYSUSPEND
+#ifdef SUSPEND_FREQ_ON
+        unsigned int sampling_factor_suspend;
+        unsigned int up_threshold_suspend;
+        unsigned int freq_step_suspend;
+        int early_suspend;
+#endif
+#endif
+} dbs_tuners_ins = {
+        .up_threshold = DEF_FREQUENCY_UP_THRESHOLD,
+        .up_threshold_min_freq = UP_THRESHOLD_AT_MIN_FREQ,
+        .sampling_down_factor = DEF_SAMPLING_DOWN_FACTOR,
+        .down_differential = DEF_FREQUENCY_DOWN_DIFFERENTIAL,
+        .ignore_nice = 0,
+        .powersave_bias = 0,
+        .freq_step = FREQ_STEP,
+        .freq_responsiveness = FREQ_FOR_RESPONSIVENESS,
+#ifdef CONFIG_HAS_EARLYSUSPEND
+#ifdef SUSPEND_FREQ_ON
+        .suspend_freq = DEF_SUSPEND_FREQ,
+        .sampling_factor_suspend = SAMPLING_FACTOR_SUSPEND,
+        .up_threshold_suspend = DEF_FREQUENCY_UP_THRESHOLD_SUSPEND,
+        .freq_step_suspend = FREQ_STEP_SUSPEND,
+        .early_suspend = -1,
+#endif
+#endif
+};
+
+static unsigned int dbs_enable=0;        /* number of CPUs using this policy */
+
+#ifdef SUSPEND_FREQ_ON
+// HYPER suspend mods (Thanks to Imoseyon)
+static unsigned int suspended = 0;
+static void HYPER_suspend(int suspend)
+{
+        struct cpu_dbs_info_s *dbs_info = &per_cpu(od_cpu_dbs_info, smp_processor_id());
+        if (dbs_enable==0) return;
+        if (!suspend) { // resume at max speed:
+                suspended = 0;
+                __cpufreq_driver_target(dbs_info->cur_policy, dbs_info->cur_policy->max,
+                        CPUFREQ_RELATION_L);
+                pr_info("[HYPER] HYPER awake at %d\n", dbs_info->cur_policy->cur);
+        } else {
+                suspended = 1;
+                // let's give it a little breathing room
+                __cpufreq_driver_target(dbs_info->cur_policy, dbs_tuners_ins.suspend_freq, CPUFREQ_RELATION_H);
+                pr_info("[HYPER] HYPER suspended at %d\n", dbs_info->cur_policy->cur);
+        }
+}
+
+static void hyper__power_early_suspend(struct early_suspend *handler) {
+       HYPER_suspend(1);
+}
+
+static void hyper__power_late_resume(struct early_suspend *handler) {
+       HYPER_suspend(0);
+}
+
+static struct early_suspend HYPER_suspend = {
+        .suspend = HYPER_early_suspend,
+        .resume = HYPER_late_resume,
+        .level = EARLY_SUSPEND_LEVEL_DISABLE_FB + 1,
+};
+#endif
+
+static inline cputime64_t get_cpu_idle_time_jiffy(unsigned int cpu,
+                                                        cputime64_t *wall)
+{
+        cputime64_t idle_time;
+        cputime64_t cur_wall_time;
+        cputime64_t busy_time;
+
+        cur_wall_time = jiffies64_to_cputime64(get_jiffies_64());
+        busy_time = cputime64_add(kstat_cpu(cpu).cpustat.user,
+                        kstat_cpu(cpu).cpustat.system);
+
+        busy_time = cputime64_add(busy_time, kstat_cpu(cpu).cpustat.irq);
+        busy_time = cputime64_add(busy_time, kstat_cpu(cpu).cpustat.softirq);
+        busy_time = cputime64_add(busy_time, kstat_cpu(cpu).cpustat.steal);
+        busy_time = cputime64_add(busy_time, kstat_cpu(cpu).cpustat.nice);
+
+        idle_time = cputime64_sub(cur_wall_time, busy_time);
+        if (wall)
+                *wall = (cputime64_t)jiffies_to_usecs(cur_wall_time);
+
+        return (cputime64_t)jiffies_to_usecs(idle_time);
+}
+
+static inline cputime64_t get_cpu_idle_time(unsigned int cpu, cputime64_t *wall)
+{
+        u64 idle_time = get_cpu_idle_time_us(cpu, NULL);
+
+        if (idle_time == -1ULL)
+                return get_cpu_idle_time_jiffy(cpu, wall);
+        else
+                idle_time += get_cpu_iowait_time_us(cpu, wall);
+
+        return idle_time;
+}
+
+static inline cputime64_t get_cpu_iowait_time(unsigned int cpu, cputime64_t *wall)
+{
+        u64 iowait_time = get_cpu_iowait_time_us(cpu, wall);
+
+        if (iowait_time == -1ULL)
+                return 0;
+
+        return iowait_time;
+}
+
+/*
+* Find right freq to be set now with powersave_bias on.
+* Returns the freq_hi to be used right now and will set freq_hi_jiffies,
+* freq_lo, and freq_lo_jiffies in percpu area for averaging freqs.
+*/
+static unsigned int powersave_bias_target(struct cpufreq_policy *policy,
+                                         unsigned int freq_next,
+                                         unsigned int relation)
+{
+        unsigned int freq_req, freq_reduc, freq_avg;
+        unsigned int freq_hi, freq_lo;
+        unsigned int index = 0;
+        unsigned int jiffies_total, jiffies_hi, jiffies_lo;
+        struct cpu_dbs_info_s *dbs_info = &per_cpu(od_cpu_dbs_info,
+                                                 policy->cpu);
+
+        if (!dbs_info->freq_table) {
+                dbs_info->freq_lo = 0;
+                dbs_info->freq_lo_jiffies = 0;
+                return freq_next;
+        }
+
+        cpufreq_frequency_table_target(policy, dbs_info->freq_table, freq_next,
+                        relation, &index);
+        freq_req = dbs_info->freq_table[index].frequency;
+        freq_reduc = freq_req * dbs_tuners_ins.powersave_bias / 1000;
+        freq_avg = freq_req - freq_reduc;
+
+        /* Find freq bounds for freq_avg in freq_table */
+        index = 0;
+        cpufreq_frequency_table_target(policy, dbs_info->freq_table, freq_avg,
+                        CPUFREQ_RELATION_H, &index);
+        freq_lo = dbs_info->freq_table[index].frequency;
+        index = 0;
+        cpufreq_frequency_table_target(policy, dbs_info->freq_table, freq_avg,
+                        CPUFREQ_RELATION_L, &index);
+        freq_hi = dbs_info->freq_table[index].frequency;
+
+        /* Find out how long we have to be in hi and lo freqs */
+        if (freq_hi == freq_lo) {
+                dbs_info->freq_lo = 0;
+                dbs_info->freq_lo_jiffies = 0;
+                return freq_lo;
+        }
+        jiffies_total = usecs_to_jiffies(dbs_tuners_ins.sampling_rate);
+        jiffies_hi = (freq_avg - freq_lo) * jiffies_total;
+        jiffies_hi += ((freq_hi - freq_lo) / 2);
+        jiffies_hi /= (freq_hi - freq_lo);
+        jiffies_lo = jiffies_total - jiffies_hi;
+        dbs_info->freq_lo = freq_lo;
+        dbs_info->freq_lo_jiffies = jiffies_lo;
+        dbs_info->freq_hi_jiffies = jiffies_hi;
+        return freq_hi;
+}
+
+static void hyper__power_powersave_bias_init_cpu(int cpu)
+{
+        struct cpu_dbs_info_s *dbs_info = &per_cpu(od_cpu_dbs_info, cpu);
+        dbs_info->freq_table = cpufreq_frequency_get_table(cpu);
+        dbs_info->freq_lo = 0;
+}
+
+static void HYPER_powersave_bias_init(void)
+{
+        int i;
+        for_each_online_cpu(i) {
+                hyper__power_powersave_bias_init_cpu(i);
+        }
+}
+
+/************************** sysfs interface ************************/
+
+static ssize_t show_sampling_rate_min(struct kobject *kobj,
+                                 struct attribute *attr, char *buf)
+{
+        return sprintf(buf, "%u\n", min_sampling_rate);
+}
+
+define_one_global_ro(sampling_rate_min);
+
+/* cpufreq_hyper__power Governor Tunables */
+#define show_one(file_name, object)                                        \
+static ssize_t show_##file_name                                                \
+(struct kobject *kobj, struct attribute *attr, char *buf) \
+{                                                                        \
+        return sprintf(buf, "%u\n", dbs_tuners_ins.object);                \
+}
+show_one(sampling_rate, sampling_rate);
+show_one(io_is_busy, io_is_busy);
+show_one(up_threshold, up_threshold);
+show_one(up_threshold_min_freq, up_threshold_min_freq);
+show_one(sampling_down_factor, sampling_down_factor);
+show_one(ignore_nice_load, ignore_nice);
+show_one(powersave_bias, powersave_bias);
+show_one(down_differential, down_differential);
+show_one(freq_step, freq_step);
+show_one(freq_responsiveness, freq_responsiveness);
+#ifdef CONFIG_HAS_EARLYSUSPEND
+#ifdef SUSPEND_FREQ_ON
+show_one(suspend_freq, suspend_freq);
+show_one(sampling_factor_suspend, sampling_factor_suspend);
+show_one(up_threshold_suspend, up_threshold_suspend);
+show_one(freq_step_suspend, freq_step_suspend);
+#endif
+#endif
+
+/**
+* update_sampling_rate - update sampling rate effective immediately if needed.
+* @new_rate: new sampling rate
+*
+* If new rate is smaller than the old, simply updaing
+* dbs_tuners_int.sampling_rate might not be appropriate. For example,
+* if the original sampling_rate was 1 second and the requested new sampling
+* rate is 10 ms because the user needs immediate reaction from ondemand
+* governor, but not sure if higher frequency will be required or not,
+* then, the governor may change the sampling rate too late; up to 1 second
+* later. Thus, if we are reducing the sampling rate, we need to make the
+* new value effective immediately.
+*/
+
+static ssize_t store_sampling_rate(struct kobject *a, struct attribute *b,
+                                 const char *buf, size_t count)
+{
+        unsigned int input;
+        int ret;
+        ret = sscanf(buf, "%u", &input);
+        if (ret != 1)
+                return -EINVAL;
+        dbs_tuners_ins.sampling_rate = max(input, min_sampling_rate);
+        return count;
+}
+
+static ssize_t store_io_is_busy(struct kobject *a, struct attribute *b,
+                                 const char *buf, size_t count)
+{
+        unsigned int input;
+        int ret;
+
+        ret = sscanf(buf, "%u", &input);
+        if (ret != 1)
+                return -EINVAL;
+
+        dbs_tuners_ins.io_is_busy = !!input;
+        return count;
+}
+
+static ssize_t store_up_threshold(struct kobject *a, struct attribute *b,
+                                 const char *buf, size_t count)
+{
+        unsigned int input;
+        int ret;
+        ret = sscanf(buf, "%u", &input);
+
+        if (ret != 1 || input > MAX_FREQUENCY_UP_THRESHOLD ||
+                        input < MIN_FREQUENCY_UP_THRESHOLD) {
+                return -EINVAL;
+        }
+        dbs_tuners_ins.up_threshold = input;
+        return count;
+}
+
+static ssize_t store_up_threshold_min_freq(struct kobject *a, struct attribute *b,
+                                 const char *buf, size_t count)
+{
+        unsigned int input;
+        int ret;
+        ret = sscanf(buf, "%u", &input);
+
+        if (ret != 1 || input > MAX_FREQUENCY_UP_THRESHOLD ||
+                        input < MIN_FREQUENCY_UP_THRESHOLD) {
+                return -EINVAL;
+        }
+        dbs_tuners_ins.up_threshold_min_freq = input;
+        return count;
+}
+
+static ssize_t store_sampling_down_factor(struct kobject *a,
+                        struct attribute *b, const char *buf, size_t count)
+{
+        unsigned int input, j;
+        int ret;
+        ret = sscanf(buf, "%u", &input);
+
+        if (ret != 1 || input > MAX_SAMPLING_DOWN_FACTOR || input < 1)
+                return -EINVAL;
+        dbs_tuners_ins.sampling_down_factor = input;
+
+        /* Reset down sampling multiplier in case it was active */
+        for_each_online_cpu(j) {
+                struct cpu_dbs_info_s *dbs_info;
+                dbs_info = &per_cpu(od_cpu_dbs_info, j);
+                dbs_info->rate_mult = 1;
+        }
+        return count;
+}
+
+static ssize_t store_ignore_nice_load(struct kobject *a, struct attribute *b,
+                                 const char *buf, size_t count)
+{
+        unsigned int input;
+        int ret;
+
+        unsigned int j;
+
+        ret = sscanf(buf, "%u", &input);
+        if (ret != 1)
+                return -EINVAL;
+
+        if (input > 1)
+                input = 1;
+
+        if (input == dbs_tuners_ins.ignore_nice) { /* nothing to do */
+                return count;
+        }
+        dbs_tuners_ins.ignore_nice = input;
+
+        /* we need to re-evaluate prev_cpu_idle */
+        for_each_online_cpu(j) {
+                struct cpu_dbs_info_s *dbs_info;
+                dbs_info = &per_cpu(od_cpu_dbs_info, j);
+                dbs_info->prev_cpu_idle = get_cpu_idle_time(j,
+                                                &dbs_info->prev_cpu_wall);
+                if (dbs_tuners_ins.ignore_nice)
+                        dbs_info->prev_cpu_nice = kstat_cpu(j).cpustat.nice;
+
+        }
+        return count;
+}
+
+static ssize_t store_powersave_bias(struct kobject *a, struct attribute *b,
+                                 const char *buf, size_t count)
+{
+        unsigned int input;
+        int ret;
+        ret = sscanf(buf, "%u", &input);
+
+        if (ret != 1)
+                return -EINVAL;
+
+        if (input > 1000)
+                input = 1000;
+
+        dbs_tuners_ins.powersave_bias = input;
+        HYPER_powersave_bias_init();
+        return count;
+}
+
+static ssize_t store_down_differential(struct kobject *a, struct attribute *b,
+                                 const char *buf, size_t count)
+{
+        unsigned int input;
+        int ret;
+        ret = sscanf(buf, "%u", &input);
+        if (ret != 1)
+                return -EINVAL;
+        dbs_tuners_ins.down_differential = min(input, 100u);
+        return count;
+}
+
+static ssize_t store_freq_step(struct kobject *a, struct attribute *b,
+                                 const char *buf, size_t count)
+{
+        unsigned int input;
+        int ret;
+        ret = sscanf(buf, "%u", &input);
+        if (ret != 1)
+                return -EINVAL;
+        dbs_tuners_ins.freq_step = min(input, 100u);
+        return count;
+}
+
+static ssize_t store_freq_responsiveness(struct kobject *a, struct attribute *b,
+                                 const char *buf, size_t count)
+{
+        unsigned int input;
+        int ret;
+        ret = sscanf(buf, "%u", &input);
+
+        if (ret != 1)
+                return -EINVAL;
+
+        if (input > 1200000)
+                input = 1200000;
+
+        if (input < 100000)
+                input = 100000;
+
+        dbs_tuners_ins.freq_responsiveness = input;
+
+        return count;
+}
+
+#ifdef CONFIG_HAS_EARLYSUSPEND
+#ifdef SUSPEND_FREQ_ON
+static ssize_t store_suspend_freq(struct kobject *a, struct attribute *b,
+                                 const char *buf, size_t count)
+{
+        unsigned int input;
+        int ret;
+        ret = sscanf(buf, "%u", &input);
+
+        if (ret != 1)
+                return -EINVAL;
+
+        if (input > 1200000)
+                input = 1200000;
+
+        if (input < 100000)
+                input = 100000;
+
+        dbs_tuners_ins.suspend_freq = input;
+
+        return count;
+}
+
+static ssize_t store_sampling_factor_suspend(struct kobject *a, struct attribute *b,
+                                 const char *buf, size_t count)
+{
+        unsigned int input;
+        int ret;
+        ret = sscanf(buf, "%u", &input);
+        if (ret != 1 || input > 10 ||
+                        input < 1) {
+                return -EINVAL;
+        }
+        dbs_tuners_ins.sampling_factor_suspend = input;
+        return count;
+}
+
+static ssize_t store_up_threshold_suspend(struct kobject *a, struct attribute *b,
+                                 const char *buf, size_t count)
+{
+        unsigned int input;
+        int ret;
+        ret = sscanf(buf, "%u", &input);
+
+        if (ret != 1 || input > MAX_FREQUENCY_UP_THRESHOLD ||
+                        input < MIN_FREQUENCY_UP_THRESHOLD) {
+                return -EINVAL;
+        }
+        dbs_tuners_ins.up_threshold_suspend = input;
+        return count;
+}
+
+static ssize_t store_freq_step_suspend(struct kobject *a, struct attribute *b,
+                                 const char *buf, size_t count)
+{
+        unsigned int input;
+        int ret;
+        ret = sscanf(buf, "%u", &input);
+        if (ret != 1)
+                return -EINVAL;
+        dbs_tuners_ins.freq_step_suspend = min(input, 100u);
+        return count;
+}
+#endif
+#endif
+
+define_one_global_rw(sampling_rate);
+define_one_global_rw(io_is_busy);
+define_one_global_rw(up_threshold);
+define_one_global_rw(up_threshold_min_freq);
+define_one_global_rw(sampling_down_factor);
+define_one_global_rw(ignore_nice_load);
+define_one_global_rw(powersave_bias);
+define_one_global_rw(down_differential);
+define_one_global_rw(freq_step);
+define_one_global_rw(freq_responsiveness);
+#ifdef CONFIG_HAS_EARLYSUSPEND
+#ifdef SUSPEND_FREQ_ON
+define_one_global_rw(suspend_freq);
+define_one_global_rw(sampling_factor_suspend);
+define_one_global_rw(up_threshold_suspend);
+define_one_global_rw(freq_step_suspend);
+#endif
+#endif
+
+static struct attribute *dbs_attributes[] = {
+        &sampling_rate_min.attr,
+        &sampling_rate.attr,
+        &up_threshold.attr,
+        &up_threshold_min_freq.attr,
+        &sampling_down_factor.attr,
+        &ignore_nice_load.attr,
+        &powersave_bias.attr,
+        &io_is_busy.attr,
+        &down_differential.attr,
+        &freq_step.attr,
+        &freq_responsiveness.attr,
+#ifdef CONFIG_HAS_EARLYSUSPEND
+#ifdef SUSPEND_FREQ_ON
+        &suspend_freq.attr,
+        &sampling_factor_suspend.attr,
+        &up_threshold_suspend.attr,
+        &freq_step_suspend.attr,
+#endif
+#endif
+        NULL
+};
+
+static struct attribute_group dbs_attr_group = {
+        .attrs = dbs_attributes,
+        .name = "HYPER",
+};
+
+/************************** sysfs end ************************/
+
+static void dbs_freq_increase(struct cpufreq_policy *p, unsigned int freq)
+{
+        if (dbs_tuners_ins.powersave_bias)
+                freq = powersave_bias_target(p, freq, CPUFREQ_RELATION_H);
+#ifndef CONFIG_ARCH_EXYNOS4
+        else if (p->cur == p->max)
+                return;
+#endif
+
+        __cpufreq_driver_target(p, freq, dbs_tuners_ins.powersave_bias ?
+                        CPUFREQ_RELATION_L : CPUFREQ_RELATION_H);
+}
+
+static void dbs_check_cpu(struct cpu_dbs_info_s *this_dbs_info)
+{
+        unsigned int max_load_freq;
+
+        struct cpufreq_policy *policy;
+        unsigned int j;
+        int up_threshold = dbs_tuners_ins.up_threshold;
+
+        this_dbs_info->freq_lo = 0;
+        policy = this_dbs_info->cur_policy;
+
+        /*
+         * Every sampling_rate, we check, if current idle time is less
+         * than 20% (default), then we try to increase frequency
+         * Every sampling_rate, we look for a the lowest
+         * frequency which can sustain the load while keeping idle time over
+         * 30%. If such a frequency exist, we try to decrease to this frequency.
+         *
+         * Any frequency increase takes it to the maximum frequency.
+         * Frequency reduction happens at minimum steps of
+         * 5% (default) of current frequency
+         */
+
+        /* Get Absolute Load - in terms of freq */
+        max_load_freq = 0;
+
+        for_each_cpu(j, policy->cpus) {
+                struct cpu_dbs_info_s *j_dbs_info;
+                cputime64_t cur_wall_time, cur_idle_time, cur_iowait_time;
+                unsigned int idle_time, wall_time, iowait_time;
+                unsigned int load, load_freq;
+                int freq_avg;
+
+                j_dbs_info = &per_cpu(od_cpu_dbs_info, j);
+
+                cur_idle_time = get_cpu_idle_time(j, &cur_wall_time);
+                cur_iowait_time = get_cpu_iowait_time(j, &cur_wall_time);
+
+                wall_time = (unsigned int) cputime64_sub(cur_wall_time,
+                                j_dbs_info->prev_cpu_wall);
+                j_dbs_info->prev_cpu_wall = cur_wall_time;
+
+                idle_time = (unsigned int) cputime64_sub(cur_idle_time,
+                                j_dbs_info->prev_cpu_idle);
+                j_dbs_info->prev_cpu_idle = cur_idle_time;
+
+                iowait_time = (unsigned int) cputime64_sub(cur_iowait_time,
+                                j_dbs_info->prev_cpu_iowait);
+                j_dbs_info->prev_cpu_iowait = cur_iowait_time;
+
+                if (dbs_tuners_ins.ignore_nice) {
+                        cputime64_t cur_nice;
+                        unsigned long cur_nice_jiffies;
+
+                        cur_nice = cputime64_sub(kstat_cpu(j).cpustat.nice,
+                                         j_dbs_info->prev_cpu_nice);
+                        /*
+                         * Assumption: nice time between sampling periods will
+                         * be less than 2^32 jiffies for 32 bit sys
+                         */
+                        cur_nice_jiffies = (unsigned long)
+                                        cputime64_to_jiffies64(cur_nice);
+
+                        j_dbs_info->prev_cpu_nice = kstat_cpu(j).cpustat.nice;
+                        idle_time += jiffies_to_usecs(cur_nice_jiffies);
+                }
+
+                /*
+                 * For the purpose of HYPER, waiting for disk IO is an
+                 * indication that you're performance critical, and not that
+                 * the system is actually idle. So subtract the iowait time
+                 * from the cpu idle time.
+                 */
+
+                if (dbs_tuners_ins.io_is_busy && idle_time >= iowait_time)
+                        idle_time -= iowait_time;
+
+                if (unlikely(!wall_time || wall_time < idle_time))
+                        continue;
+
+                load = 100 * (wall_time - idle_time) / wall_time;
+
+                freq_avg = __cpufreq_driver_getavg(policy, j);
+                if (freq_avg <= 0)
+                        freq_avg = policy->cur;
+
+                load_freq = load * freq_avg;
+                if (load_freq > max_load_freq)
+                        max_load_freq = load_freq;
+        }
+
+        /* Check for frequency increase */
+#ifdef SUSPEND_FREQ_ON
+        if (policy->cur < dbs_tuners_ins.freq_responsiveness
+                && dbs_tuners_ins.early_suspend == -1) {
+                        up_threshold = dbs_tuners_ins.up_threshold_min_freq;
+        }
+#else
+        if (policy->cur < dbs_tuners_ins.freq_responsiveness) {
+                        up_threshold = dbs_tuners_ins.up_threshold_min_freq;
+        }
+#endif
+        if (max_load_freq > up_threshold * policy->cur) {
+                int inc = (policy->max * dbs_tuners_ins.freq_step) / 100;
+                int target = min(policy->max, policy->cur + inc);
+                /* If switching to max speed, apply sampling_down_factor */
+                if (policy->cur < policy->max && target == policy->max)
+                        this_dbs_info->rate_mult =
+                                dbs_tuners_ins.sampling_down_factor;
+                dbs_freq_increase(policy, target);
+                return;
+        }
+
+        /* Check for frequency decrease */
+#ifndef CONFIG_ARCH_EXYNOS4
+        /* if we cannot reduce the frequency anymore, break out early */
+        if (policy->cur == policy->min)
+                return;
+#endif
+
+        /*
+         * The optimal frequency is the frequency that is the lowest that
+         * can support the current CPU usage without triggering the up
+         * policy. To be safe, we focus 10 points under the threshold.
+         */
+        if (max_load_freq <
+         (dbs_tuners_ins.up_threshold - dbs_tuners_ins.down_differential) *
+         policy->cur) {
+                unsigned int freq_next;
+                unsigned int down_thres;
+                freq_next = max_load_freq /
+                                (dbs_tuners_ins.up_threshold -
+                                 dbs_tuners_ins.down_differential);
+
+                /* No longer fully busy, reset rate_mult */
+                this_dbs_info->rate_mult = 1;
+
+                if (freq_next < policy->min)
+                        freq_next = policy->min;
+
+                down_thres = dbs_tuners_ins.up_threshold_min_freq
+                        - dbs_tuners_ins.down_differential;
+
+#ifdef SUSPEND_FREQ_ON
+                if (freq_next < dbs_tuners_ins.freq_responsiveness
+                        && (max_load_freq / freq_next) > down_thres
+                                && dbs_tuners_ins.early_suspend == -1)
+                        freq_next = dbs_tuners_ins.freq_responsiveness;
+#else
+                if (freq_next < dbs_tuners_ins.freq_responsiveness
+                        && (max_load_freq / freq_next) > down_thres)
+                        freq_next = dbs_tuners_ins.freq_responsiveness;
+#endif
+
+                if (!dbs_tuners_ins.powersave_bias) {
+                        __cpufreq_driver_target(policy, freq_next,
+                                        CPUFREQ_RELATION_L);
+                } else {
+                        int freq = powersave_bias_target(policy, freq_next,
+                                        CPUFREQ_RELATION_L);
+                        __cpufreq_driver_target(policy, freq,
+                                CPUFREQ_RELATION_L);
+                }
+        }
+}
+
+static void do_dbs_timer(struct work_struct *work)
+{
+        struct cpu_dbs_info_s *dbs_info =
+                container_of(work, struct cpu_dbs_info_s, work.work);
+        unsigned int cpu = dbs_info->cpu;
+        int sample_type = dbs_info->sample_type;
+
+        int delay;
+
+        mutex_lock(&dbs_info->timer_mutex);
+
+        /* Common NORMAL_SAMPLE setup */
+        dbs_info->sample_type = DBS_NORMAL_SAMPLE;
+        if (!dbs_tuners_ins.powersave_bias ||
+         sample_type == DBS_NORMAL_SAMPLE) {
+                dbs_check_cpu(dbs_info);
+                if (dbs_info->freq_lo) {
+                        /* Setup timer for SUB_SAMPLE */
+                        dbs_info->sample_type = DBS_SUB_SAMPLE;
+                        delay = dbs_info->freq_hi_jiffies;
+                } else {
+                        /* We want all CPUs to do sampling nearly on
+                         * same jiffy
+                         */
+                        delay = usecs_to_jiffies(dbs_tuners_ins.sampling_rate
+                                * dbs_info->rate_mult);
+
+                        if (num_online_cpus() > 1)
+                                delay -= jiffies % delay;
+                }
+        } else {
+#ifdef SUSPEND_FREQ_ON
+                if (!suspended)
+#endif
+                __cpufreq_driver_target(dbs_info->cur_policy,
+                        dbs_info->freq_lo, CPUFREQ_RELATION_H);
+                delay = dbs_info->freq_lo_jiffies;
+        }
+        schedule_delayed_work_on(cpu, &dbs_info->work, delay);
+        mutex_unlock(&dbs_info->timer_mutex);
+}
+
+static inline void dbs_timer_init(struct cpu_dbs_info_s *dbs_info)
+{
+        /* We want all CPUs to do sampling nearly on same jiffy */
+        int delay = usecs_to_jiffies(dbs_tuners_ins.sampling_rate);
+
+        if (num_online_cpus() > 1)
+                delay -= jiffies % delay;
+
+        dbs_info->sample_type = DBS_NORMAL_SAMPLE;
+        INIT_DELAYED_WORK_DEFERRABLE(&dbs_info->work, do_dbs_timer);
+        schedule_delayed_work_on(dbs_info->cpu, &dbs_info->work, delay);
+}
+
+static inline void dbs_timer_exit(struct cpu_dbs_info_s *dbs_info)
+{
+        cancel_delayed_work_sync(&dbs_info->work);
+}
+
+/*
+* Not all CPUs want IO time to be accounted as busy; this dependson how
+* efficient idling at a higher frequency/voltage is.
+* Pavel Machek says this is not so for various generations of AMD and old
+* Intel systems.
+* Mike Chan (androidlcom) calis this is also not true for ARM.
+* Because of this, whitelist specific known (series) of CPUs by default, and
+* leave all others up to the user.
+*/
+static int should_io_be_busy(void)
+{
+#if defined(CONFIG_X86)
+        /*
+         * For Intel, Core 2 (model 15) andl later have an efficient idle.
+         */
+        if (boot_cpu_data.x86_vendor == X86_VENDOR_INTEL &&
+         boot_cpu_data.x86 == 6 &&
+         boot_cpu_data.x86_model >= 15)
+                return 1;
+#endif
+        return 0;
+}
+
+#ifdef CONFIG_HAS_EARLYSUSPEND
+#ifdef SUSPEND_FREQ_ON
+static struct early_suspend early_suspend;
+unsigned int prev_up_threshold_HYPER;
+unsigned int prev_freq_step_HYPER;
+unsigned int prev_sampling_rate_HYPER;
+static void cpufreq_HYPER_early_suspend(struct early_suspend *h)
+{
+        dbs_tuners_ins.early_suspend = 1;
+        prev_freq_step_HYPER = dbs_tuners_ins.freq_step;
+        prev_sampling_rate_HYPER = dbs_tuners_ins.sampling_rate;
+        prev_up_threshold_HYPER = dbs_tuners_ins.up_threshold;
+        dbs_tuners_ins.freq_step = dbs_tuners_ins.freq_step_suspend;
+        dbs_tuners_ins.sampling_rate *= dbs_tuners_ins.sampling_factor_suspend;
+        dbs_tuners_ins.up_threshold = dbs_tuners_ins.up_threshold_suspend;
+}
+static void cpufreq_HYPER_late_resume(struct early_suspend *h)
+{
+        dbs_tuners_ins.early_suspend = -1;
+        dbs_tuners_ins.freq_step = prev_freq_step_HYPER;
+        dbs_tuners_ins.sampling_rate = prev_sampling_rate_HYPER;
+        dbs_tuners_ins.up_threshold = prev_up_threshold_HYPER;
+}
+#endif
+#endif
+
+static int cpufreq_governor_dbs(struct cpufreq_policy *policy,
+                                 unsigned int event)
+{
+        unsigned int cpu = policy->cpu;
+        struct cpu_dbs_info_s *this_dbs_info;
+        unsigned int j;
+        int rc;
+
+        this_dbs_info = &per_cpu(od_cpu_dbs_info, cpu);
+
+        switch (event) {
+        case CPUFREQ_GOV_START:
+                if ((!cpu_online(cpu)) || (!policy->cur))
+                        return -EINVAL;
+
+                mutex_lock(&dbs_mutex);
+
+                dbs_enable++;
+                for_each_cpu(j, policy->cpus) {
+                        struct cpu_dbs_info_s *j_dbs_info;
+                        j_dbs_info = &per_cpu(od_cpu_dbs_info, j);
+                        j_dbs_info->cur_policy = policy;
+
+                        j_dbs_info->prev_cpu_idle = get_cpu_idle_time(j,
+                                                &j_dbs_info->prev_cpu_wall);
+                        if (dbs_tuners_ins.ignore_nice) {
+                                j_dbs_info->prev_cpu_nice =
+                                                kstat_cpu(j).cpustat.nice;
+                        }
+                }
+                this_dbs_info->cpu = cpu;
+                this_dbs_info->rate_mult = 1;
+                hyper__power_powersave_bias_init_cpu(cpu);
+                /*
+                 * Start the timerschedule work, when this governor
+                 * is used for first time
+                 */
+                if (dbs_enable == 1) {
+                        unsigned int latency;
+
+                        rc = sysfs_create_group(cpufreq_global_kobject,
+                                                &dbs_attr_group);
+                        if (rc) {
+                                mutex_unlock(&dbs_mutex);
+                                return rc;
+                        }
+
+                        /* policy latency is in nS. Convert it to uS first */
+                        latency = policy->cpuinfo.transition_latency / 1000;
+                        if (latency == 0)
+                                latency = 1;
+                        /* Bring kernel and HW constraints together */
+                        min_sampling_rate = max(min_sampling_rate,
+                                        MIN_LATENCY_MULTIPLIER * latency);
+                        dbs_tuners_ins.sampling_rate =
+                                max(min_sampling_rate,
+                                 latency * LATENCY_MULTIPLIER);
+                        dbs_tuners_ins.io_is_busy = should_io_be_busy();
+                }
+                mutex_unlock(&dbs_mutex);
+
+                mutex_init(&this_dbs_info->timer_mutex);
+                dbs_timer_init(this_dbs_info);
+#ifdef CONFIG_HAS_EARLYSUSPEND
+#ifdef SUSPEND_FREQ_ON
+                register_early_suspend(&early_suspend);
+                register_early_suspend(&HYPER_power_suspend);
+#endif
+#endif
+                break;
+
+        case CPUFREQ_GOV_STOP:
+#ifdef CONFIG_HAS_EARLYSUSPEND
+#ifdef SUSPEND_FREQ_ON
+                unregister_early_suspend(&early_suspend);
+                unregister_early_suspend(&HYPER_power_suspend);
+#endif
+#endif
+                dbs_timer_exit(this_dbs_info);
+
+                mutex_lock(&dbs_mutex);
+                mutex_destroy(&this_dbs_info->timer_mutex);
+                dbs_enable--;
+                mutex_unlock(&dbs_mutex);
+                if (!dbs_enable)
+                        sysfs_remove_group(cpufreq_global_kobject,
+                                         &dbs_attr_group);
+                break;
+
+        case CPUFREQ_GOV_LIMITS:
+                mutex_lock(&this_dbs_info->timer_mutex);
+                if (policy->max < this_dbs_info->cur_policy->cur)
+                        __cpufreq_driver_target(this_dbs_info->cur_policy,
+                                policy->max, CPUFREQ_RELATION_H);
+                else if (policy->min > this_dbs_info->cur_policy->cur)
+                        __cpufreq_driver_target(this_dbs_info->cur_policy,
+                                policy->min, CPUFREQ_RELATION_L);
+                mutex_unlock(&this_dbs_info->timer_mutex);
+                break;
+        }
+        return 0;
+}
+
+static int __init cpufreq_gov_dbs_init(void)
+{
+        u64 idle_time;
+        int cpu = get_cpu();
+
+        idle_time = get_cpu_idle_time_us(cpu, NULL);
+        put_cpu();
+        if (idle_time != -1ULL) {
+                /* Idle micro accounting is supported. Use finer thresholds */
+                dbs_tuners_ins.up_threshold = MICRO_FREQUENCY_UP_THRESHOLD;
+                dbs_tuners_ins.down_differential =
+                                        MICRO_FREQUENCY_DOWN_DIFFERENTIAL;
+                /*
+                 * In no_hz/micro accounting case we set the minimum frequency
+                 * not depending on HZ, but fixed (very low). The deferred
+                 * timer might skip some samples if idle/sleeping as needed.
+                */
+                min_sampling_rate = MICRO_FREQUENCY_MIN_SAMPLE_RATE;
+        } else {
+                /* For correct statistics, we need 10 ticks for each measure */
+                min_sampling_rate =
+                        MIN_SAMPLING_RATE_RATIO * jiffies_to_usecs(1);
+        }
+
+#ifdef CONFIG_HAS_EARLYSUSPEND
+#ifdef SUSPEND_FREQ_ON
+        early_suspend.level = EARLY_SUSPEND_LEVEL_DISABLE_FB;
+        early_suspend.suspend = cpufreq_HYPER_early_suspend;
+        early_suspend.resume = cpufreq_HYPER_late_resume;
+#endif
+#endif
+
+        return cpufreq_register_governor(&cpufreq_gov_HYPER);
+}
+
+static void __exit cpufreq_gov_dbs_exit(void)
+{
+        cpufreq_unregister_governor(&cpufreq_gov_HYPER);
+}
+
+MODULE_AUTHOR("Venkatesh Pallipadi <venkatesh.pallipadi@intel.com>");
+MODULE_AUTHOR("Alexey Starikovskiy <alexey.y.starikovskiy@intel.com>");
+MODULE_AUTHOR("Dorimanx <yuri@bynet.co.il>");
+MODULE_DESCRIPTION("'cpufreq_HYPER' - A dynamic cpufreq governor for "
+        "Low Latency Frequency Transition capable processors"
+"Module include, IOWAIT,FREQ-TUNERS,EARLY-SUSPEND,SUSPEND-FREQ");
+MODULE_LICENSE("GPL");
+
+#ifdef CONFIG_CPU_FREQ_DEFAULT_GOV_HYPER
+fs_initcall(cpufreq_gov_dbs_init);
+#else
+module_init(cpufreq_gov_dbs_init);
+#endif
+module_exit(cpufreq_gov_dbs_exit);
